requirements.txt
anthropic==0.74.1
Babel==2.17.0
boto3==1.40.27
botocore==1.40.27
caseflow-agents==1.0.0
Flask==3.1.2
GitPython==3.1.45
gunicorn==23.0.0
mistralai==1.9.11
networkx==3.4.2
openai==2.8.1
phonenumbers==9.0.13
pydantic==2.12.4
python-dotenv==1.2.1
requests==2.32.5
sofia-utilities==1.0.0
thefuzz==0.22.1
tiktoken==0.9.0
transitions==0.9.3
